requirements.txt
Flask==2.3.3
gunicorn==21.2.0
supabase==2.18.1
httpx==0.24.1
flask-cors==4.0.0
werkzeug==2.3.7
python-dotenv==1.0.0